# docs/openapi/inventory-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Inventory Service API
  description: API documentation for ACM Inventory Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/public/lookup/inventory/receive-harvest:
    post:
      tags:
      - public-internal-lookup-controller
      operationId: receiveFromHarvest
      parameters:
      - name: harvestId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: actorUserId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveHarvestRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductWarehouseLotDto"
  /api/v1/public/lookup/inventory/lots/{lotId}/sync:
    post:
      tags:
      - public-internal-lookup-controller
      operationId: syncLinkedLotFromHarvest
      parameters:
      - name: lotId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncLotRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductWarehouseLotDto"
...
  /api/v1/public/lookup/supplies/validate-lot:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: validateSupplyLot
      parameters:
      - name: lotId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: itemId
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: farmIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ValidationResultDto"
  /api/v1/public/lookup/supplies/validate-item:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: validateSupplyItem
      parameters:
      - name: itemId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: farmIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ValidationResultDto"
  /api/v1/public/lookup/supplies/lots/{id}/batch-code:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: getSupplyLotBatchCode
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/public/lookup/supplies/items/{id}/name:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: getSupplyItemName
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/public/lookup/inventory/stock-context:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: findHarvestStockContext
      parameters:
      - name: farmId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: warehouseId
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: productName
        in: query
        required: true
        schema:
          type: string
      - name: lotCode
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HarvestStockContextDto"
  /api/v1/public/lookup/inventory/lots/by-seasons:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: findLotsBySeasonIds
      parameters:
      - name: seasonIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWarehouseLotDto"
  /api/v1/public/lookup/inventory/lots/by-harvests:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: findLotsByHarvestIds
      parameters:
      - name: harvestIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWarehouseLotDto"
  /api/v1/public/lookup/inventory/lots/by-harvest/{harvestId}:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: findLotByHarvestId
      parameters:
      - name: harvestId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductWarehouseLotDto"
  /api/v1/public/lookup/inventory/exists-by-harvest/{harvestId}:
    get:
      tags:
      - public-internal-lookup-controller
      operationId: existsProductWarehouseLotByHarvestId
      parameters:
      - name: harvestId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
...
components:
  schemas:
...
    ReceiveHarvestRequest:
      type: object
      properties:
        seasonId:
          type: integer
          format: int32
        seasonName:
          type: string
        farmId:
          type: integer
          format: int32
        farmName:
          type: string
        plotId:
          type: integer
          format: int32
        plotName:
          type: string
        cropId:
          type: integer
          format: int32
        cropName:
          type: string
        varietyId:
          type: integer
          format: int32
        varietyName:
          type: string
        harvestDate:
          type: string
          format: date
        quantity:
          type: number
        grade:
          type: string
        warehouseId:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int32
        productId:
          type: integer
          format: int32
        productName:
          type: string
        productVariant:
          type: string
        lotCode:
          type: string
        unit:
          type: string
        note:
          type: string
    ProductWarehouseLotDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        lotCode:
          type: string
        productId:
          type: integer
          format: int32
        productName:
          type: string
        productVariant:
          type: string
        harvestId:
          type: integer
          format: int32
        warehouseId:
          type: integer
          format: int32
        locationId:
          type: integer
          format: int32
        harvestedAt:
          type: string
          format: date
        unit:
          type: string
        initialQuantity:
          type: number
        onHandQuantity:
          type: number
        grade:
          type: string
        qualityStatus:
          type: string
        note:
          type: string
        status:
          type: string
          enum:
          - IN_STOCK
          - HOLD
          - DEPLETED
          - ARCHIVED
    SyncLotRequest:
      type: object
      properties:
        harvestedAt:
          type: string
          format: date
        initialQuantity:
          type: number
        onHandQuantity:
          type: number
        grade:
          type: string
        qualityStatus:
          type: string
        note:
          type: string
        status:
          type: string
...
    ValidationResultDto:
      type: object
      properties:
        valid:
          type: boolean
        errorCode:
          type: string
        errorMessage:
          type: string
        resolvedItemId:
          type: integer
          format: int32
    HarvestStockContextDto:
      type: object
      properties:
        warehouseName:
          type: string
        matchingLots:
          type: integer
          format: int64
        onHandQuantity:
          type: number
        unit:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT